+++ packages/core/docs/.vuepress/config.js
@@ -9,7 +9,8 @@ module.exports = {
 		nav: [
 			{ text: 'Home', link: '/' },
 			{ text: 'Demo', link: 'https://vsf-next-demo.storefrontcloud.io' },
-			{ text: 'Migration guide', link: '/migrate/' }
+			{ text: 'Migration guide', link: '/migrate/' },
+			{ text: 'Roadmap', link: 'https://www.notion.so/vuestorefront/Vue-Storefront-2-Next-High-level-Roadmap-201cf06abb314b84ad01b7b8463c0437'}
 
 		],
 		sidebar: {
@@ -17,9 +18,17 @@ module.exports = {
 				{
 					title: 'Migration guide RC.1',
 					children: [
-						['/migrate/', 'Overview'],
+						['/migrate/rc1', 'Overview'],
 						['/migrate/integrators-rc1', 'Integrators'],
 						['/migrate/projects-rc1', 'Projects'],
+					]
+				},
+				{
+					title: 'Migration guide 2.2.0',
+					children: [
+						['/migrate/2.2.0', 'Overview'],
+						['/migrate/integrators-2.2.0', 'Integrators'],
+						['/migrate/projects-2.2.0', 'Projects'],
 					]
 				}
 			],
@@ -138,6 +147,7 @@ module.exports = {
 					collapsable: false,
 					children: [
 						['/guide/theme', 'Theme'],
+						['/guide/configuration', 'Configuration'],
 						['/guide/error-handling', 'Error Handling']
 					]
 				},
@@ -187,8 +197,7 @@ module.exports = {
 					collapsable: true,
 					children: [
 						['/integrate/integration-guide', 'Integration guide'],
-						['/general/cms', 'CMS'],
-						['/integrate/api-middleware', 'API middleware'],
+						['/general/cms', 'CMS']
 					]
 				},
 				{